pyproject.toml
[tool.poetry]
authors = ["Arash Hatami <hatamiarash7@gmail.com>"]
classifiers = [
  "Topic :: Software Development",
  "Topic :: Software Development :: Libraries",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Development Status :: 5 - Production/Stable",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "Framework :: FastAPI",
  "Typing :: Typed",
]
description = "Export Prometheus metrics for Iran football league"
homepage = "https://arash-hatami.ir"
keywords = ["iran", "football", "prometheus", "exporter", "metrics", "monitoring"]
license = "MIT"
name = "iranleague-exporter"
packages = [{include = "iranleague_exporter"}]
readme = "README.md"
repository = "https://github.com/hatamiarash7/iranleague-exporter"
version = "1.2.0"

[tool.poetry.dependencies]
beautifulsoup4 = "^4.12.3"
fastapi = "^0.123.0"
jdatetime = "^5.0.0"
prometheus-client = "^0.23.0"
python = "^3.10"
python-slugify = "^8.0.4"
requests = "^2.32.3"
toml = "^0.10.2"
urllib3 = "^2.0.0"
uvicorn = "^0.38.0"

[tool.poetry.scripts]
start = "iranleague_exporter.main:start"

[tool.poetry.group.test.dependencies]
httpx = "^0.28.1"
pytest = "^9.0.0"
pytest-asyncio = "^1.0.0"
pytest-cov = "^7.0.0"
responses = "^0.25.0"

[tool.poetry.group.dev.dependencies]
black = "^25.0.0"
flake8-pyproject = "^1.2.3"
isort = "^7.0.0"
mypy = "^1.13.0"
pylint = "^4.0.4"
pytest = "^9.0.0"
pytest-asyncio = "^1.0.0"
ruff = "^0.8.0"
types-requests = "^2.32.0"
types-toml = "^0.10.8"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/hatamiarash7/iranleague-exporter/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.flake8]
per-file-ignores = ["test_*.py:E501"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]
indent-width = 4
line-length = 88
target-version = "py311"

[tool.ruff.analyze]
detect-string-imports = true

[tool.ruff.lint]
extend-select = [
  'D', # pydocstyle
  'E',
  'W', # pycodestyle
  'F', # pyflakes
  'I', # sort imports
  'UP', # pyupgrade
  "RUF", # ruff dev's own rules
  "SIM", # pyflakes simplicity
  "C90", # more complexity rules
]

[tool.ruff.lint.per-file-ignores]
"**/{tests,docs,tools}/*" = ["E402", "D101"]
"__init__.py" = ["E402"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = false

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = "dynamic"
exclude = ["*.pyi"]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
python_version = "3.10"
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "jdatetime.*",
  "slugify.*",
  "bs4.*",
  "prometheus_client.*",
]
